Value Added 21 October
This problem has been asked several times in many TOP MNC's.
Test Case:
5
8 10 12 16 21
For this test case, elements are 8 12 10 16 21. as 8, 12 16 are divisible by 4 so for 1st element is 8, will be the root 12: divisible by 4, so added at left of root 10 not divisible, added at right of root 16 divisible, root->left , but it is not null, again new root is 12 , added left of 12 (sub root) 21 not divisible , root->right , but it is not null, again new root is 10, added right of 10 (sub root of sub tree) 8 12 10 16 21 so level order traversal is 8 12 10 16 21.