Compute the Result
The first input will be a single integer N denoting the number of test cases to take. After this there will be exactly N lines, each line a valid postfix string. The string will be a valid postfix expression consisting of only integers and binary operators (+, -, *, / and ^).

Every integer and operator will be compulsorily separated by a SPACE. The symbol ‘^’ denotes the exponent operator.

Input Format

The first input will be a single integer N denoting the number of test cases to take. After this there will be exactly N lines, each line a valid postfix string. The string will be a valid postfix expression consisting of only integers and binary operators (+, -, *, / and ^).

Every integer and operator will be compulsorily separated by a SPACE. The symbol ‘^’ denotes the exponent operator.

Constraints

1 <= |postfix_string| <= 100

Output Format

Exactly N lines, each line denoting the output of the expression. In case the output is in fractions, please print only the integer part of the output.

Sample Input 0

2
7 5 2 + * 4 1 5 - / -
5 6 2 ^ 2 - *
Sample Output 0

50
170